Ispravlen 2 commit

+++ src/main/java/ru/netology/javaqa/javaqamvn/service/StatsService.java
@@ -2,24 +2,24 @@ package ru.netology.javaqa.javaqamvn.service;
 
 public class StatsService {
 
-    public int getSum(int[] sales) {
+    public long getSum(int[] sales) {
 
-        int findSum = 0;
+        long findSum = 0;
         for (int i = 0; i < sales.length; i++) {
             findSum = findSum + sales[i];
         }
         return findSum; // Сумма всех продаж
     }
 
-    public int getMediumSum(int[] sales) {
+    public long getMediumSum(int[] sales) {
 
-        int findSum = getSum(sales);
-        int month = sales.length;
-        int findMediumSum = findSum / sales.length;
+        long findSum = getSum(sales);
+        long month = sales.length;
+        long findMediumSum = findSum / sales.length;
         return findMediumSum; // Средняя сумма продаж в месяц
     }
 
-    public int getMaxMonth(int[] sales) {
+    public long getMaxMonth(int[] sales) {
 
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
@@ -30,7 +30,7 @@ public class StatsService {
         return maxMonth + 1; // Номер месяца, в котором был пик продаж, то есть осуществлены продажи на максимальную сумму
     }
 
-    public int getMinMonth(int[] sales) {
+    public long getMinMonth(int[] sales) {
 
         int minMonth = 0;
         for (int i = 0; i < sales.length; i++) {
@@ -41,10 +41,10 @@ public class StatsService {
         return minMonth + 1; // Номер месяца, в котором был минимум продаж, то есть осуществлены продажи на минимальную сумму
     }
 
-    public int monthsBelowMedium(int[] sales) {
+    public long monthsBelowMedium(int[] sales) {
 
-        int findMediumSum = getMediumSum(sales);
-        int monthBelowMedium = 0;
+        long findMediumSum = getMediumSum(sales);
+        long monthBelowMedium = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] < findMediumSum) {
                 monthBelowMedium = monthBelowMedium + 1;
@@ -53,10 +53,10 @@ public class StatsService {
         return (monthBelowMedium); // Количество месяцев, в которых продажи были ниже среднего
     }
 
-    public int monthsHigherMedium(int[] sales) {
+    public long monthsHigherMedium(int[] sales) {
 
-        int findMediumSum = getMediumSum(sales);
-        int monthsHigherMedium = 0;
+        long findMediumSum = getMediumSum(sales);
+        long monthsHigherMedium = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] > findMediumSum) {
                 monthsHigherMedium = monthsHigherMedium + 1;
