+++ src/test/java/ru/netology/javaqa/javaqamvn/service/StatsServiceTest.java
@@ -12,8 +12,8 @@ public class StatsServiceTest {
                 16, 42, 86, 14, 55, 67, 11, 142, 17, 64, 5, 19
         };
 
-        int expected = 538;
-        int actual = service.getSum(sales);
+        long expected = 538;
+        long actual = service.getSum(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -24,8 +24,8 @@ public class StatsServiceTest {
                 16, 42, 86, 14, 55, 67, 11, 142, 17, 64, 5, 19
         };
 
-        int expected = 44;
-        int actual = service.getMediumSum(sales);
+        long expected = 44;
+        long actual = service.getMediumSum(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -36,8 +36,8 @@ public class StatsServiceTest {
                 16, 42, 86, 14, 55, 67, 11, 142, 17, 64, 5, 19
         };
 
-        int expected = 8;
-        int actual = service.getMaxMonth(sales);
+        long expected = 8;
+        long actual = service.getMaxMonth(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -48,8 +48,8 @@ public class StatsServiceTest {
                 16, 42, 86, 14, 55, 67, 11, 142, 17, 64, 5, 19
         };
 
-        int expected = 11;
-        int actual = service.getMinMonth(sales);
+        long expected = 11;
+        long actual = service.getMinMonth(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -60,8 +60,8 @@ public class StatsServiceTest {
                 16, 42, 86, 14, 55, 67, 11, 142, 17, 64, 5, 19
         };
 
-        int expected = 7;
-        int actual = service.monthsBelowMedium(sales);
+        long expected = 7;
+        long actual = service.monthsBelowMedium(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -72,8 +72,8 @@ public class StatsServiceTest {
                 16, 42, 86, 14, 55, 67, 11, 142, 17, 64, 5, 19
         };
 
-        int expected = 5;
-        int actual = service.monthsHigherMedium(sales);
+        long expected = 5;
+        long actual = service.monthsHigherMedium(sales);
         Assertions.assertEquals(expected, actual);
     }
 }
